Separate out the application from the server script
